rename with-next-js to with-nextjs-postgres
