+++ javascript/with-nextjs-postgres/package.json
@@ -11,6 +11,7 @@
     "next": "9.5.4",
     "pg": "^8.4.0",
     "react": "16.13.1",
-    "react-dom": "16.13.1"
+    "react-dom": "16.13.1",
+    "swr": "^0.3.9"
   }
 }
