+++ javascript/with-nextjs-postgres/pages/api/time.js
@@ -1,7 +1,6 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 
-// Call Railway anywhere on serverside code.
-// Here or getInitialProps
+// Use Railway Postgres anywhere in the API routes or in getServerSideProps
 
 import pg from "pg";
 
@@ -12,9 +11,7 @@ export default async (req, res) => {
     const query = await client.query("SELECT NOW()");
 
     res.json({
-      data: {
-        time: query.rows[0].now,
-      },
+      time: query.rows[0].now,
     });
   } catch (e) {
     console.log("ERROR", e);
