+++ javascript/with-nextjs-postgres/public/favicon.ico
